Refactor code by splitting app.js components #6
